Fix the wait-read loops
Channel#read_stream catches EAGAIN errors, so the blocking_call wrapper was
useless. On the other hand, the loop must perform a waitsocket to compensate
for the lack of waiting from read_stream. Otherwise, the loops made the CPU run
at full throttle.

The native call to libssh2_channel_read_ex passed the size of the pointer
instead of the size of the buffer to libssh2, making it always return tiny
amounts of data. The allocated buffer was also never freed, so I use the stack
instead.